refactor: resolve #2462 — extract ConstructionLayer, PerSurfaceConductionSolver, SurfaceKind into fluxion-core
